requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
grpcio==1.46.3
grpcio-tools==1.46.3
idna==3.4
music-tag==0.4.3
mutagen==1.46.0
protobuf==3.20.3
pydub==0.25.1
requests==2.31.0
six==1.16.0
urllib3==2.0.4
yandex-speechkit==1.3.3
